+++ src/Main.java
@@ -1,37 +1,48 @@
+import com.flight.system.Flight;
+import com.flight.system.Flights;
+import com.flight.system.ReadScheduler;
+import com.flight.system.util.Validator;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.util.List;
+import java.util.Locale;
 
 public class Main {
 
     BufferedReader reader;
+    Flights flights;
 
     public Main() {
         reader = new BufferedReader(new InputStreamReader(System.in));
+        flights = new Flights();
     }
-    Students students = new Students();
 
     public static void main(String[] args) throws IOException {
         Main obj = new Main();
+        ReadScheduler rs = new ReadScheduler(obj.flights);
+
         boolean running = true;
         System.out.println("What would you like to perform!");
-        while(running){
-            System.out.println("1. Add a student");
-            System.out.println("2. Print Students");
+        while (running) {
+            System.out.println("1. Search flights");
+            System.out.println("2. Print all flights");
             System.out.println("3. Exit");
-            switch(obj.reader.readLine()){
+            switch (obj.reader.readLine()) {
 
                 case "1":
-                    obj.readStudentDetails();
+                    obj.getQuery();
                     break;
 
                 case "2":
-                    obj.printStudents();
+                    obj.printAllFlights();
                     break;
 
                 case "3":
                     System.out.println("Thank you");
                     running = false;
+                    rs.shutdown();
                     break;
 
                 default:
@@ -39,23 +50,32 @@ public class Main {
                     break;
             }
         }
-        obj.reader.close();
     }
 
-    public void readStudentDetails() throws IOException {
-        System.out.println("Enter Student Name!");
-        String name = this.reader.readLine();
-        System.out.println("Enter Student Age!");
-        int age = Integer.parseInt(this.reader.readLine());
-        System.out.println("Enter Father's Name!");
-        String fathersName = this.reader.readLine();
-        Student record = new Student(age, name, fathersName);
-        if(!students.addStudent(record)) {
-            System.out.println("duplicate!!");
+    public void getQuery() throws IOException {
+        System.out.println("Enter Departure Location!");
+        String depLoc = this.reader.readLine().toUpperCase(Locale.ROOT);
+        System.out.println("Enter Arrival Location!");
+        String arrLoc = this.reader.readLine().toUpperCase(Locale.ROOT);
+        System.out.println("Enter Flight Date (dd-mm-yyyy)!");
+        String flightDate = this.reader.readLine().toUpperCase(Locale.ROOT);
+        if (!Validator.isThreeLetterCode.test(depLoc) || !Validator.isThreeLetterCode.test(arrLoc) || !Validator.isValidDate.test(flightDate)) {
+            System.out.println("Invalid input!!");
+            return;
         }
+        List<Flight> result = flights.searchFlights(depLoc, arrLoc, flightDate);
+        if (result.isEmpty()) {
+            System.out.println("Sorry, no flights available for your query!!\n");
+        }
+        result.stream()
+                .forEach(System.out::println);
     }
 
-    public void printStudents() {
-        System.out.println(students.toString());
+    public void printAllFlights() {
+        flights.getAllFlights()
+                .stream()
+                .forEach(System.out::println);
+        System.out.println(flights.getTotalFlights());
     }
+
 }
