+++ src/com/flight/system/Flight.java
@@ -1,5 +1,7 @@
 package com.flight.system;
 
+import com.flight.system.util.Utilities;
+
 import java.text.DecimalFormat;
 import java.util.Date;
