+++ src/com/flight/system/Flights.java
@@ -1,5 +1,7 @@
 package com.flight.system;
 
+import com.flight.system.util.Utilities;
+
 import java.util.*;
 import java.util.stream.Collectors;
